updated tests, refactor
